Organize React components
