+++ src/components/Blog.tsx
@@ -1,9 +1,5 @@
 import React from 'react'
-
-const posts = [
-  { id: 1, title: 'Welcome to My Blog', content: 'This is a space where I share my thoughts and experiences.' },
-  { id: 2, title: 'First Post', content: 'Feel free to explore and connect with me!' },
-]
+import { posts } from '../data/posts'
 
 function Blog() {
   return (
@@ -15,6 +11,9 @@ function Blog() {
           <p>{post.content}</p>
         </article>
       ))}
+      <p>
+        Read more on my <a href="/blog/">Astro blog</a>.
+      </p>
     </div>
   )
 }
