+++ src/components/Home.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import './Home.css'
+
+function Home() {
+  return (
+    <div className="home-container">
+      <h1>Welcome to My Blog</h1>
+      <p>This is a space where I share my thoughts and experiences.</p>
+      <p>Feel free to explore and connect with me!</p>
+    </div>
+  )
+}
+
+export default Home
